# docs/swagger.yaml
basePath: /
definitions:
  masterBalance.CreateDTO:
    properties:
      from_id:
        type: string
      money_amount:
        type: string
      order_id:
        type: string
      service_id:
        type: string
    type: object
  user.User:
    properties:
      balance:
        type: string
      id:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: dimvas2010@yandex.ru
  description: This service help to work with users balance
  title: Avito backend internship
  version: "1.0"
paths:
  /billing:
    post:
      consumes:
      - application/json
      description: Add money to user's balance with billing systems (visa/mastercard)
      parameters:
      - description: Users Info
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Add money to user's balance
      tags:
      - accounts
      - billings
  /moneyAccept:
    post:
      consumes:
      - application/json
      description: Accept money from master balance when service is done
      parameters:
      - description: Request to freeze money
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/masterBalance.CreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Accepts money
      tags:
      - accounts
      - reserve
  /moneyFreeze:
    post:
      consumes:
      - application/json
      description: Reserves money from user balance to special master account
      parameters:
      - description: Request to freeze money
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/masterBalance.CreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Reserves money
      tags:
      - accounts
      - reserve
  /moneyReject:
    post:
      consumes:
      - application/json
      description: Return money to user when payment for service is rejected
      parameters:
      - description: Request to freeze money
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/masterBalance.CreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Rejects money
      tags:
      - accounts
      - reserve
      - reject
  /report/{month}/{year}:
    get:
      consumes:
      - application/json
      description: Return link to report.csv file with money for every service
      parameters:
      - description: Needed month for report
        in: path
        name: month
        required: true
        type: integer
      - description: Needed year for report
        in: path
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Returns report for date range
      tags:
      - accounts
      - balance
      - report
  /transactions/{userid}/{pageNum}/{sortSum}/{sortDate}:
    get:
      consumes:
      - application/json
      description: Return text with history of transactions
      parameters:
      - description: id of needed user
        in: path
        name: userid
        required: true
        type: integer
      - description: number of searching page
        in: path
        name: pageNum
        required: true
        type: integer
      - description: Parameter for sort by sum (asc/desc)
        in: path
        name: sortSum
        required: true
        type: string
      - description: Parameter for sort by date (asc/desc)
        in: path
        name: sortDate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Returns info about user transactions
      tags:
      - accounts
      - balance
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Return user account with his balance
      parameters:
      - description: User's id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Returns user balance
      tags:
      - accounts
      - balance
swagger: "2.0"